requirements.txt
exceptiongroup==1.1.1
Faker==18.5.1
greenlet==2.0.2
grpcio==1.54.0
grpcio-testing==1.54.0
grpcio-tools==1.54.0
iniconfig==2.0.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
pluggy==1.0.0
protobuf==4.22.3
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
PyJWT==2.6.0
pylama==8.4.1
pytest==7.3.1
python-dateutil==2.8.2
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==2.0.11
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.3
